+++ lib/src/realtime/realtime.dart
@@ -1,8 +1,0 @@
-library;
-
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'realtime.freezed.dart';
-part 'realtime.g.dart';
-
-part 'src/message.dart';